[FIX] Add compiler defines to identify PCIe design
The PCIe design does not support VEth interface currently.
Due to this the openPOWERLINK kernel library for dualprocshm
interface should be compiled without the VEth.

As the dualprocshm is used with other design such as Zynq, a
compiler define is added to identify the PCIe design and disable
the VEth only for PCIe designs.

Change-Id: Idfaa35bd8bc10359f9571fc7ead8318c91ec16a7
Signed-off-by: Gaurav Kumar Singh <gaurav@kalycito.com>